type casting added for predict_csv

+++ api/main.py
@@ -41,7 +41,7 @@ def predict(input: PredictRequest, model: Model = Depends(get_model)):
 @app.post("/predict_csv")
 def predict_csv(csv_file: UploadFile = File(...), model: Model = Depends(get_model)):
     try:
-        df = pd.read_csv(csv_file.file)
+        df = pd.read_csv(csv_file.file).astype(float)
     except:
         raise HTTPException(
             status_code=HTTP_422_UNPROCESSABLE_ENTITY, detail="Unable to process file"
